Clamp values and user float for current

+++ migrations/000024_recommendation_sets_add_current_request_columns.up.sql
@@ -2,5 +2,5 @@
 -- Existing rows remain NULL; the poller fills values for new records.
 -- List queries use ORDER BY ... DESC NULLS LAST (see listoptions.SQLOrderByFragment).
 ALTER TABLE recommendation_sets
-    ADD COLUMN cpu_request_current NUMERIC(10, 4),
-    ADD COLUMN memory_request_current NUMERIC(10, 4);
+    ADD COLUMN cpu_request_current FLOAT,
+    ADD COLUMN memory_request_current FLOAT;
